fase 1 lista eso creo

+++ src/main/java/com/miguelangel/taqueria/services/TacoService.java
@@ -29,7 +29,13 @@ public class TacoService {
         return tacoRepository.save(taco);
     }
 
-    public void eliminar(Integer id) {
+    public boolean eliminar(Integer id) {
+
+        if (!tacoRepository.existsById(id)) {
+            return false;
+        }
+
         tacoRepository.deleteById(id);
+        return true;
     }
 }
